Aislación de partes escenciales
Se separó el footer y la barra superior en distintos archivos y fragments, con tal de poder exportarlos a páginas secundarias. Se cambiaron los nombres de algunos archivos para hacer más claro su contenido.
